Provide recognized variants of ohsumed and reuters
