+++ ludwig/datasets/reuters_cmu/__init__.py
@@ -21,17 +21,17 @@ from ludwig.datasets.registry import register_dataset
 
 
 def load(cache_dir=DEFAULT_CACHE_LOCATION, split=False):
-    dataset = OhsuMed(cache_dir=cache_dir)
+    dataset = Reuters(cache_dir=cache_dir)
     return dataset.load(split=split)
 
 
-@register_dataset(name="ohsumed")
-class OhsuMed(ZipDownloadMixin, IdentityProcessMixin, CSVLoadMixin, BaseDataset):
-    """The OhsuMed dataset.
+@register_dataset(name="reuters_cmu")
+class Reuters(ZipDownloadMixin, IdentityProcessMixin, CSVLoadMixin, BaseDataset):
+    """The Reuters dataset.
 
     This pulls in an array of mixins for different types of functionality which belongs in the workflow for ingesting
     and transforming training data into a destination dataframe that can fit into Ludwig's training API.
     """
 
     def __init__(self, cache_dir=DEFAULT_CACHE_LOCATION):
-        super().__init__(dataset_name="ohsumed", cache_dir=cache_dir)
+        super().__init__(dataset_name="reuters_cmu", cache_dir=cache_dir)
