+++ ludwig/datasets/reuters_r8/__init__.py
@@ -1,5 +1,5 @@
 #! /usr/bin/env python
-# Copyright (c) 2019 Uber Technologies, Inc.
+# Copyright (c) 2021 Uber Technologies, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,25 +13,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
+
 from ludwig.datasets.base_dataset import BaseDataset, DEFAULT_CACHE_LOCATION
-from ludwig.datasets.mixins.download import ZipDownloadMixin
+from ludwig.datasets.mixins.kaggle import KaggleDownloadMixin
 from ludwig.datasets.mixins.load import CSVLoadMixin
-from ludwig.datasets.mixins.process import IdentityProcessMixin
+from ludwig.datasets.mixins.process import MultifileJoinProcessMixin
 from ludwig.datasets.registry import register_dataset
 
 
-def load(cache_dir=DEFAULT_CACHE_LOCATION, split=False):
-    dataset = Reuters(cache_dir=cache_dir)
+def load(cache_dir=DEFAULT_CACHE_LOCATION, split=False, kaggle_username=None, kaggle_key=None):
+    dataset = ReutersR8(cache_dir=cache_dir, kaggle_username=kaggle_username, kaggle_key=kaggle_key)
     return dataset.load(split=split)
 
 
-@register_dataset(name="reuters")
-class Reuters(ZipDownloadMixin, IdentityProcessMixin, CSVLoadMixin, BaseDataset):
-    """The Reuters dataset.
+@register_dataset(name="reuters_r8")
+class ReutersR8(CSVLoadMixin, MultifileJoinProcessMixin, KaggleDownloadMixin, BaseDataset):
+    """Reuters R8 subset of Reuters 21578 dataset from Kaggle."""
 
-    This pulls in an array of mixins for different types of functionality which belongs in the workflow for ingesting
-    and transforming training data into a destination dataframe that can fit into Ludwig's training API.
-    """
-
-    def __init__(self, cache_dir=DEFAULT_CACHE_LOCATION):
-        super().__init__(dataset_name="reuters", cache_dir=cache_dir)
+    def __init__(self, cache_dir=DEFAULT_CACHE_LOCATION, kaggle_username=None, kaggle_key=None):
+        self.kaggle_username = kaggle_username
+        self.kaggle_key = kaggle_key
+        self.is_kaggle_competition = False
+        super().__init__(dataset_name="reuters_r8", cache_dir=cache_dir)
